Fix long and short desc
